+++ Cargo.toml
@@ -12,6 +12,7 @@ path = "src/lib.rs"
 [dependencies]
 actix-web = { version = "4", features = ["macros"] }
 async-trait = { version = "0.1" }
+backtrace = { version = "0.3.76", optional = true }
 chrono = { version = "0.4.42" }
 clap = { version = "4", features = ["derive"] }
 dotenv = { version = "0.15" }
@@ -21,7 +22,7 @@ once_cell = { version = "1.21.3" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0" }
 serde_yaml = { version = "0.9.33" }
-shaku = { version = "0.6.2" , features = ["derive"] }
+shaku = { version = "0.6.2", features = ["derive"] }
 sqlx = { version = "0.8", features = ["runtime-tokio", "sqlite"] }
 strfmt = { version = "0.2.5" }
 tokio = { version = "1", features = ["full"] }
@@ -30,8 +31,11 @@ utoipa = { version = "5.4.0" }
 utoipa-swagger-ui = { version = "9.0.2", features = ["actix-web"] }
 uuid = { version = "1", features = ["v4", "serde"] }
 
+[features]
+test-utils = ["backtrace"]
+
 [dev-dependencies]
-backtrace = "0.3.76"
+ebisu_api = { path = ".", features = ["test-utils"] }
 sqlx-cli = { version = "0.8" }
 
 [build-dependencies]
