+++ src/lib.rs
@@ -2,3 +2,8 @@ pub mod dao; // DAO層
 pub mod entity; // エンティティ層
 pub mod handler; // ハンドラー層
 pub mod utils; // ユーティリティ層
+
+#[cfg(any(test, feature = "test-utils"))]
+pub mod test_helpers {
+    pub use crate::utils::unit_test::fixtures::*;
+}
