+++ src/utils/unit_test/fixtures/app_test_setup.rs
@@ -1,4 +1,3 @@
-#![cfg(test)]
 use crate::handler::accounts::AccountHandlerImpl;
 use crate::utils::unit_test::fixtures::accounts::ParametrizedMockAccountDaoImpl;
 use shaku::module;
