+++ src/utils/unit_test/fixtures/db.rs
@@ -1,4 +1,4 @@
-#[cfg(test)]
+use crate::utils::unit_test::fixtures::accounts;
 use sqlx::{Executor, SqlitePool, sqlite::SqlitePoolOptions};
 use std::fs;
 use uuid::Uuid;
@@ -6,42 +6,41 @@ use uuid::Uuid;
 const MIGRATIONS_DIR: &str = "./migrations";
 
 /// テスト用のインメモリSQLiteデータベースをセットアップするヘルパー関数
-///
-/// # Returns:
-///   SqlitePool - マイグレーションが適用されたSQLite接続プール
-/// # Panics
-///   マイグレーションの適用に失敗した場合、パニックします。
+
+// 空のデータベースを作成する
+pub async fn create_empty_db() -> SqlitePool {
+    db_migration().await
+}
+
+// テストデータが入ったデータベースを作成する
 pub async fn create_test_db() -> SqlitePool {
-    let db_name = format!("file:memdb-{}", Uuid::new_v4().to_string());
-    let db_url = format!("{}?mode=memory&cache=shared", db_name);
-    let pool = SqlitePoolOptions::new().connect(&db_url).await.unwrap();
-    let mut conn = pool.acquire().await.unwrap();
-    // マイグレーションファイルを読み込み、順番に実行する
-    let mut migration_paths: Vec<_> = fs::read_dir(MIGRATIONS_DIR)
-        .expect("Failed to read migrations directory")
-        .filter_map(|entry| {
-            let path = entry.ok()?.path();
-            if path.extension().map_or(false, |ext| ext == "sql") { Some(path) } else { None }
-        })
-        .collect();
-    migration_paths.sort();
-    for path in migration_paths {
-        let sql_content = fs::read_to_string(&path).expect(&format!("Failed to read SQL file: {}", path.display()));
-        conn.execute(sql_content.as_str())
-            .await
-            .unwrap_or_else(|e| panic!("Failed to execute migration: {}: {}", path.display(), e));
-    }
+    let pool = create_empty_db().await;
+    accounts::insert_test_account(&pool).await;
     pool
 }
 
-/// スキーマが設定されていないDBへのアクセスプールを作成する
-/// # Returns:
-///   SqlitePool - スキーマが設定されていないSQLite接続プール
-/// # Panics
-///   接続プールの作成に失敗した場合、パニックします。
+// スキーマが設定されていないDBへのアクセスプールを作成する
 pub async fn create_undefined_db() -> SqlitePool {
     let db_name = format!("file:memdb-{}", Uuid::new_v4().to_string());
     let db_url = format!("{}?mode=memory&cache=shared", db_name);
-    let pool = SqlitePoolOptions::new().max_connections(1).connect(&db_url).await.unwrap();
+    let pool = SqlitePoolOptions::new().connect(&db_url).await.unwrap();
+    pool
+}
+
+// マイグレーションを適用したテスト用データベースを作成する
+pub async fn db_migration() -> SqlitePool {
+    let pool = create_undefined_db().await;
+    let mut conn = pool.acquire().await.unwrap();
+    // マイグレーションファイルを読み込み、順番に実行する
+    let paths = fs::read_dir(MIGRATIONS_DIR).expect("Failed to read migrations directory");
+    for entry in paths {
+        let path = entry.expect("Failed to get entry").path();
+        if path.extension().map_or(false, |ext| ext == "sql") {
+            let sql_content = fs::read_to_string(&path).expect(&format!("Failed to read SQL file: {}", path.display()));
+            conn.execute(sql_content.as_str())
+                .await
+                .expect(&format!("Failed to execute migration: {}", path.display()));
+        }
+    }
     pool
 }
