+++ src/utils/unit_test/fixtures/mod.rs
@@ -1,4 +1,3 @@
-#![cfg(test)]
 pub mod accounts; // アカウント関連のfixture
 pub mod app_test_setup; // アプリテスト設定関連のfixture
 pub mod db; // テストDB関連のfixture
