+++ src/utils/unit_test/mod.rs
@@ -1,2 +1,1 @@
-#[cfg(test)]
 pub mod fixtures; // fixture関連
